Changed all uses of OwnedArray<File> to be Array<File> instead, which is a bit more efficient and easier to work with. Optimised the software renderer to improve most drawing operations, particularly simple shapes filling like drawing lines, etc.

+++ src/gui/components/special/juce_MagnifierComponent.cpp
@@ -275,7 +275,9 @@ void MagnifierComponent::paint (Graphics& g)
     }
 
     g.setImageResamplingQuality (quality);
-    g.drawImageTransformed (&temp, temp.getBounds(), AffineTransform::scale (scaleFactor, scaleFactor), false);
+    g.drawImageTransformed (&temp, temp.getBounds(),
+                            AffineTransform::scale ((float) scaleFactor, (float) scaleFactor),
+                            false);
 }
 
 void MagnifierComponent::childBoundsChanged (Component* c)
